@@ -16,7 +16,7 @@
 
 package org.cloudfoundry.tomcat.logging.access;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.PrintStream;
 
