@@ -17,10 +17,10 @@
 package org.cloudfoundry.tomcat.lifecycle;
 
 import org.apache.catalina.LifecycleState;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public final class StandardStateParserTest {
 
